LayoutTests/imported/w3c/web-platform-tests/wasm/jsapi/memory/type.tentative.any-expected.txt


FAIL Zero initial, no maximum assert_equals: expected (number) 0 but got (undefined) undefined
FAIL Non-zero initial, no maximum assert_equals: expected (number) 5 but got (undefined) undefined
FAIL Zero maximum assert_equals: expected (number) 0 but got (undefined) undefined
FAIL None-zero maximum assert_equals: expected (number) 0 but got (undefined) undefined
FAIL non-shared memory assert_equals: expected (number) 0 but got (undefined) undefined
FAIL shared memory assert_equals: expected (number) 0 but got (undefined) undefined
